+++ packages/cdtn-types/index.ts
@@ -111,3 +111,27 @@ export type EditorialContentDataWrapper = {
   anchor: string[];
   information: EditorialContentData;
 };
+
+export type Tool = {
+  date: string;
+  icon: string;
+  order: number;
+  action: string;
+  metaTitle?: string;
+  questions?: string[];
+  description: string;
+  displayTitle?: string;
+  breadcrumbs: Record<string, string | number>[];
+  cdtnId: string;
+  excludeFromSearch: boolean;
+  id: string;
+  isPublished: boolean;
+  metaDescription: string;
+  slug: string;
+  source: string;
+  text: string;
+  title: string;
+  title_vector: number[];
+  _id: string;
+  disable?: boolean;
+};
