+++ packages/code-du-travail-frontend/__tests__/outils.test.tsx
@@ -1,4 +1,4 @@
-import { render } from "@testing-library/react";
+import { prettyDOM, render } from "@testing-library/react";
 import React from "react";
 
 import Outils, { getServerSideProps } from "../pages/outils";
@@ -12,14 +12,12 @@ describe("<Outils />", () => {
     serverSideProps = await getServerSideProps();
   });
 
-  it("should render", async () => {
-    const { container } = render(<Outils {...serverSideProps.props} />);
-    expect(container).toBeTruthy();
-  });
-
   it("should display a list of tools", () => {
-    const { getByTestId } = render(<Outils {...serverSideProps.props} />);
-    const grid = getByTestId("tools-list");
-    expect(grid.getElementsByTagName("a")).toHaveLength(11);
+    const { queryAllByTestId, debug } = render(
+      <Outils {...serverSideProps.props} />
+    );
+    debug();
+    const items = queryAllByTestId("tools-list-items-internal");
+    expect(items).toHaveLength(2);
   });
 });
