+++ packages/code-du-travail-frontend/src/outils/__mocks__/service.ts
@@ -1,7 +1,8 @@
-import { getToolsParams, Tool } from "../service";
+import { Tool } from "cdtn-types";
+import { getToolsParams } from "../service";
 
 export const fetchTools = async (_params: getToolsParams): Promise<Tool[]> => {
-  return [
+  return Promise.resolve([
     {
       date: "18/10/2018",
       icon: "Indemnity",
@@ -83,277 +84,5 @@ export const fetchTools = async (_params: getToolsParams): Promise<Tool[]> => {
       _id: "1eea193273",
       title_vector: [],
     },
-    {
-      date: "14/03/2019",
-      icon: "Salary",
-      order: 4,
-      action: "Estimer",
-      metaTitle: "Simulateur - Calcul du salaire brut/net",
-      questions: [
-        "Comment calculer mon salaire",
-        "calculer mon salaire",
-        "calcul salaire",
-        "montant de mon salaire",
-        "montant du salaire",
-        "montant salaire",
-        "montant salaire net",
-        "montant salaire brut",
-        "calcul du salaire net",
-        "net en brut",
-        "brut en net",
-        "calcul net en brut",
-        "calcul brut en net",
-        "montant net en brut",
-        "montant brut en net",
-        "cout embauche CDD",
-        "cout embauche CDI",
-      ],
-      description:
-        "Estimez le salaire : brut en net, net après impôt, calcul heures supplémentaires, coût total employeur",
-      displayTitle: "Calculer le salaire brut/net",
-      breadcrumbs: [
-        {
-          label: "Salaire et Rémunération",
-          position: 1,
-          slug: "/themes/salaire-et-remuneration",
-        },
-        { label: "Salaire", position: 0, slug: "/themes/salaire" },
-      ],
-      cdtnId: "d7ad36850a",
-      excludeFromSearch: false,
-      id: "4718de1b-dda5-47e7-a797-c14e62cf7259",
-      isPublished: true,
-      metaDescription:
-        "Estimer le salaire : brut en net, net après impôt, calcul heures supplémentaires, coût total employeur",
-      slug: "simulateur-embauche",
-      source: "outils",
-      text: "Comment calculer mon salaire\ncalculer mon salaire\ncalcul salaire\nmontant de mon salaire\nmontant du salaire\nmontant salaire\nmontant salaire net\nmontant salaire brut\ncalcul du salaire net\nnet en brut\nbrut en net\ncalcul net en brut\ncalcul brut en net\nmontant net en brut\nmontant brut en net\ncout embauche CDD\ncout embauche CDI",
-      title: "Salaire brut/net",
-      _id: "d7ad36850a",
-      title_vector: [],
-    },
-    {
-      date: "14/03/2019",
-      icon: "Precarity",
-      order: 5,
-      action: "Estimer",
-      metaTitle: "Simulateur - Calcul de l'indemnité de précarité",
-      questions: [
-        "fin de contrat cdd",
-        "indemnité fin de contrat",
-        "indemnité cdd",
-        "calcul de la prime de précarité",
-        "Comment se calcul la prime de précarité",
-        "montant prime de précarité",
-        "montant indemnité fin de contrat",
-      ],
-      description: "Estimez le montant de l’indemnité de fin de CDD ou intérim",
-      displayTitle: "Calculer l'indemnité de précarité",
-      breadcrumbs: [
-        {
-          label: "Départ de l’entreprise",
-          position: 7,
-          slug: "/themes/depart-de-lentreprise",
-        },
-        {
-          label: "Fin d’un CDD - CTT",
-          position: 2,
-          slug: "/themes/fin-dun-cdd-ctt",
-        },
-      ],
-      cdtnId: "65a8c4a452",
-      excludeFromSearch: false,
-      id: "1f092661-e948-4f27-aafd-7f5142f25f46",
-      isPublished: true,
-      metaDescription:
-        "Estimez le montant de l’indemnité de fin de CDD ou intérim",
-      slug: "indemnite-precarite",
-      source: "outils",
-      text: "fin de contrat cdd\nprime fin de contrat\nprime de précarité\nindemnité fin de contrat\nindemnité cdd\ncalcul de la prime de précarité\nComment se calcul la prime de précarité\nmontant prime de précarité\nmontant indemnité fin de contrat",
-      title: "Indemnité de précarité",
-      _id: "65a8c4a452",
-      title_vector: [],
-    },
-    {
-      date: "08/11/2019",
-      icon: "Termination",
-      order: 6,
-      action: "Calculer",
-      metaTitle: "Simulateur - Calcul du préavis de licenciement",
-      questions: [
-        "preavis de licenciement",
-        "preavis de licenciement CDI",
-        "durée du préavis de licenciement",
-        "calculer la durée de son préavis de licenciement",
-        "calcul durée préavis licenciement",
-      ],
-      description:
-        "Estimez la durée de préavis à respecter en cas de licenciement",
-      displayTitle: "Calculer le préavis de licenciement",
-      breadcrumbs: [
-        {
-          label: "Départ de l’entreprise",
-          position: 7,
-          slug: "/themes/depart-de-lentreprise",
-        },
-        {
-          label: "Licenciement : droits des salariés et procédures",
-          position: 4,
-          slug: "/themes/licenciement-droits-des-salaries-et-procedures",
-        },
-      ],
-      cdtnId: "2fac5f4163",
-      excludeFromSearch: false,
-      id: "fcc43645-9ac1-433d-bd95-0eb44239b864",
-      isPublished: true,
-      metaDescription:
-        "Calculez la durée de préavis à respecter en cas de licenciement",
-      slug: "preavis-licenciement",
-      source: "outils",
-      text: "preavis de licenciement\npreavis de licenciement CDI\ndurée du préavis de licenciement\ncalculer la durée de son préavis de licenciement\ncalcul durée préavis licenciement",
-      title: "Préavis de licenciement",
-      _id: "2fac5f4163",
-      title_vector: [],
-    },
-    {
-      date: "08/11/2019",
-      icon: "Time",
-      order: 7,
-      action: "Calculer",
-      metaTitle:
-        "Simulateur - Calcul du nombre d’heures d'absence pour rechercher un emploi",
-      questions: [
-        "ai-je droit à des heures de recherche d'emploi",
-        "quel est le nombre d'heures pour rechercher un emploi",
-        "rechercher un emploi en période de préavis",
-        "comment prendre ses heures de recherche d'emploi",
-        "démission recherche emploi pendant préavis",
-        "licenciement recherche d emploi pendant préavis",
-        "heures de recherche d'emploi en cas de licenciement",
-        "Ai-je droit aux heures pour recherche d'emploi dans le cadre d'une démission",
-        "autorisation d absence pour recherche d emploi",
-        "chercher un emploi pendant son préavis",
-        "absence 2h pour recherche d'emploi pendant préavis",
-      ],
-      description:
-        "Estimez le nombre d'heures d'absence autorisée pour rechercher un emploi pendant un préavis",
-      displayTitle:
-        "Calculer le nombre d'heures d'absence pour rechercher un emploi",
-      breadcrumbs: [
-        {
-          label: "Départ de l’entreprise",
-          position: 7,
-          slug: "/themes/depart-de-lentreprise",
-        },
-      ],
-      cdtnId: "d8fca710c5",
-      excludeFromSearch: false,
-      id: "d09665fc-961c-44bf-a399-c34cfedf1d24",
-      isPublished: true,
-      metaDescription:
-        "Calculez le nombre d’heures pour recherche d’emploi prévues pendant le préavis",
-      slug: "heures-recherche-emploi",
-      source: "outils",
-      text: "ai-je droit à des heures de recherche d’emploi\nquel est le nombre d’heures pour rechercher un emploi\nrechercher un emploi en période de préavis\ncomment prendre ses heures de recherche d’emploi\ndémission recherche emploi pendant préavis\nlicenciement recherche d emploi pendant préavis\nheures de recherche d’emploi en cas de licenciement\nAi-je droit aux heures pour recherche d’emploi dans le cadre d’une démission\nautorisation d absence pour recherche d emploi\nchercher un emploi pendant son préavis\nabsence 2h pour recherche d’emploi pendant préavis",
-      title: "Heures d'absence pour rechercher un emploi",
-      _id: "d8fca710c5",
-      title_vector: [],
-    },
-    {
-      date: "01/01/2020",
-      icon: "SearchCC",
-      order: 8,
-      action: "Consulter",
-      metaTitle: "Trouver sa convention collective",
-      questions: [
-        "trouver ma convention collective",
-        "convention collective",
-        "recherche ccn",
-        "trouver ma cc",
-      ],
-      description:
-        "Recherchez une convention collective par Entreprise, SIRET, Nom ou numéro IDCC",
-      displayTitle: "Trouver sa convention collective",
-      breadcrumbs: [
-        {
-          label: "Représentation du personnel et négociation collective",
-          position: 6,
-          slug: "/themes/representation-du-personnel-et-negociation-collective",
-        },
-        {
-          label: "Négociations de branche",
-          position: 4,
-          slug: "/themes/negociations-de-branche",
-        },
-      ],
-      cdtnId: "db8ffe3574",
-      excludeFromSearch: false,
-      id: "c249bd2a-6f33-4176-ad09-166bb8e5747b",
-      isPublished: true,
-      metaDescription:
-        "Recherchez une convention collective par Entreprise, SIRET, Nom ou numéro IDCC",
-      slug: "convention-collective",
-      source: "outils",
-      text: "trouver ma convention collective\nconvention collective\nrecherche ccn\ntrouver ma cc",
-      title: "Trouver sa convention collective",
-      _id: "db8ffe3574",
-      title_vector: [],
-    },
-    {
-      date: "18/11/2021",
-      icon: "CalendarTime",
-      order: 9,
-      action: "Calculer",
-      metaTitle:
-        "Simulateur - Calcul du préavis de départ ou mise à la retraite",
-      description:
-        "Estimez la durée de préavis à respecter en cas de départ à la retraite ou de mise à la retraite",
-      displayTitle: "Calculer le préavis de départ à la retraite",
-      breadcrumbs: [
-        {
-          label: "Départ de l’entreprise",
-          position: 7,
-          slug: "/themes/depart-de-lentreprise",
-        },
-        { label: "Retraite", position: 5, slug: "/themes/retraite" },
-      ],
-      cdtnId: "1e518ff4e2",
-      excludeFromSearch: false,
-      id: "9b54c48e-134f-48a7-8924-cad91168357e",
-      isPublished: true,
-      metaDescription:
-        "Calculez la durée de préavis à respecter en cas de départ à la retraite ou de mise à la retraite",
-      slug: "preavis-retraite",
-      source: "outils",
-      text: "Calculez la durée de préavis à respecter en cas de départ à la retraite ou de mise à la retraite",
-      title: "Préavis de départ ou de mise à la retraite",
-      title_vector: [],
-      _id: "1e518ff4e2",
-      questions: [],
-    },
-    {
-      date: "05/07/2022",
-      icon: "FileCheck",
-      order: 10,
-      action: "Consulter",
-      metaTitle: "Simulateur - Comprendre sa procédure de licenciement",
-      questions: [],
-      description:
-        "Accédez à la procédure de licenciement qui correspond à votre situation en répondant à quelques questions",
-      displayTitle: "Comprendre sa procédure de licenciement",
-      breadcrumbs: [],
-      cdtnId: "7bd7f6defa",
-      excludeFromSearch: false,
-      id: "74b1c67b-f3c4-479a-a112-7f6739a6ece5",
-      isPublished: true,
-      metaDescription:
-        "Accédez à la procédure de licenciement qui correspond à votre situation en répondant à quelques questions",
-      slug: "procedure-licenciement",
-      source: "outils",
-      text: "Accédez à la procédure de licenciement qui correspond à votre situation en répondant à quelques questions",
-      title: "Comprendre sa procédure de licenciement",
-      title_vector: [],
-      _id: "7bd7f6defa",
-    },
-  ];
+  ]);
 };
